# backend/internal/extension/schema/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Amitia Extension Runtime API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/extensions/packages": {
      "get": {
        "responses": {
          "200": {
            "description": "Installed extension package list"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "List installed extension packages"
      }
    },
    "/extensions/packages/artifacts": {
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {"format": "binary", "type": "string"},
                  "scopeId": {"type": "string"},
                  "scopeType": {"enum": ["global", "character"], "type": "string"}
                },
                "required": ["file", "scopeType"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {"description": "Artifact stored and preview session created"},
          "422": {"description": "Package rejected by the verification pipeline"}
        },
        "summary": "Upload an immutable package artifact and create its preview"
      }
    },
    "/extensions/packages/previews/{sessionId}/confirm": {
      "post": {
        "parameters": [{"in": "path", "name": "sessionId", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"content": {"application/json": {"schema": {"type": "object"}}}, "required": true},
        "responses": {
          "200": {"description": "Bound, expiring confirmation token"},
          "422": {"description": "Preview or confirmations rejected"}
        },
        "summary": "Confirm a package preview"
      }
    },
...
    "/extensions/packages/operations/install": {
      "post": {
        "requestBody": {"content": {"application/json": {"schema": {"type": "object"}}}, "required": true},
        "responses": {
          "202": {"description": "Install operation completed and journaled"},
          "422": {"description": "Confirmation, policy, artifact, or trust validation failed"}
        },
        "summary": "Execute a confirmed package install operation"
      }
    },
    "/extensions/packages/operations/update": {
      "post": {
        "requestBody": {"content": {"application/json": {"schema": {"type": "object"}}}, "required": true},
        "responses": {
          "202": {"description": "Update operation completed and journaled"},
          "422": {"description": "Confirmation, policy, artifact, dependency, or update validation failed"}
        },
        "summary": "Execute a confirmed package update operation"
      }
    },
    "/extensions/packages/operations/{operationId}": {
      "get": {
        "parameters": [{"in": "path", "name": "operationId", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Authoritative operation journal"}, "404": {"description": "Operation not found"}},
        "summary": "Get a package operation"
      }
    },
...
    "/extensions/packages/install": {
      "post": {
        "deprecated": true,
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "package": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "package"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "410": {
            "description": "PACKAGE_INSTALL_ENDPOINT_RETIRED"
          }
        },
        "summary": "Retired package install endpoint",
        "x-replacement-endpoint": "/api/extensions/packages/artifacts"
      }
    },
...
    "/extensions/packages/status": {
      "get": {
        "responses": {
          "200": {
            "description": "Extension package status"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get extension package service status"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/api"
    }
  ]
}